+++ pages/export.php
@@ -8,7 +8,7 @@ require_once(__ROOT__.'/inc/settings.php');
 
 if($_GET['format'] == 'csv' && $_GET['kind'] == 'ingredients'){
 	$defCatClass = $settings['defCatClass'];
-	$r = mysqli_query($conn, "SELECT name,INCI,cas,FEMA,type,strength,category,profile,$defCatClass,physical_state,allergen FROM ingredients");
+	$r = mysqli_query($conn, "SELECT name,INCI,cas,FEMA,type,strength,profile,physical_state,allergen,odor,impact_top,impact_heart,impact_base FROM ingredients");
 	
 	$ing = array();
 	if (mysqli_num_rows($r) > 0) {
@@ -20,7 +20,7 @@ if($_GET['format'] == 'csv' && $_GET['kind'] == 'ingredients'){
 	header('Content-Type: text/csv; charset=utf-8');
 	header('Content-Disposition: attachment; filename='.$_GET['kind'].'.csv');
 	$output = fopen('php://output', 'w');
-	fputcsv($output, array('Name', 'INCI', 'CAS', 'FEMA', 'Type', 'Strength', 'Category', 'Profile', 'Category Class', 'Physical State', 'Allergen'));
+	fputcsv($output, array('Name', 'INCI', 'CAS', 'FEMA', 'Type', 'Strength', 'Profile', 'Physical State', 'Allergen', 'Odor Description', 'Top Note Impact', 'Heart Note Impact', 'Base Note Impact'));
 	
 	if (count($ing) > 0) {
 		foreach ($ing as $row) {