+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 - Set database to connection to utf-8 by default
 - Filter IUPAC in ingredients for illegal chars
 - Improvements in SDS document generation
+- Ingredients CSV import/export extended to support allergen status and notes impact
+- Minor improvements in CSV import
 
 ### Version 6.2
 - Fix a bug giving an javascript error when you creating a new empty formula